Moving error_from_status to a class method of RecurlyError
