+++ recurly/__init__.py
@@ -21,20 +21,7 @@ USER_AGENT = "Recurly/{}; python {}; {}".format(
 STRICT_MODE = os.getenv("RECURLY_STRICT_MODE", "FALSE").upper() == "TRUE"
 
 
-class RecurlyError(Exception):
-    pass
-
-
-class ApiError(RecurlyError):
-    def __init__(self, message, error):
-        super(ApiError, self).__init__(message)
-        self.error = error
-
-
-class NetworkError(RecurlyError):
-    pass
-
-
+from .base_errors import RecurlyError, ApiError, NetworkError
 from . import errors
 from .client import Client
 from .resource import Resource
